feat: refactoring middleware
