+++ applications/network/lib/middleware/index.ts
@@ -1,6 +1,3 @@
-export { default as Context } from "./context";
-export { default as State } from "./state";
-
 export { default as pingPongMiddleware } from "./ping-pong.middleware";
 export { default as versionMiddleware } from "./version.middleware";
