+++ applications/network/lib/middleware/ping-pong.middleware.ts
@@ -1,6 +1,6 @@
 import Application from "koa";
-import Context from "./context";
-import State from "./state";
+import Context from "../context";
+import State from "../state";
 
 const pingPongMiddleware: Application.Middleware<State, Context> = async (
   context,
