+++ applications/network/lib/routes/ping-pong.routes.ts
@@ -1,6 +1,7 @@
 import Router from "koa-router";
 
-import { Context, pingPongMiddleware } from "../middleware";
+import Context from "../context";
+import { pingPongMiddleware } from "../middleware";
 
 function pigPongRoutes(): Router<never, Context> {
   const router = new Router<never, Context>();
