+++ applications/network/lib/state.ts
